+++ goyek.ps1
@@ -1,6 +1,6 @@
 $ErrorActionPreference = "Stop"
 
 Push-Location "$PSScriptRoot\build"
-& go run . $args
+& go run . -wd=".." $args
 Pop-Location
 exit $global:LASTEXITCODE